refactor navigation our of App.tsx into MainNavigation.tsx
