+++ navigation/Routes.ts
@@ -1,0 +1,7 @@
+export const Routes = {
+  PetTypes: 'PetTypes',
+  Breeds: 'Breeds',
+  Animals: 'Animals',
+  Options: 'Options',
+  AnimalDetails: 'AnimalDetails',
+};
